+++ tests/integration/Requests/PublicGroupRequestTest.php
@@ -4,7 +4,6 @@
 namespace Tests\Integration\Requests;
 
 
-use CTApi\Models\GroupCategory;
 use CTApi\Models\GroupInformation;
 use CTApi\Models\TargetGroup;
 use CTApi\Requests\PublicGroupRequest;
@@ -45,9 +44,6 @@ class PublicGroupRequestTest extends TestCaseAuthenticated
 
         $this->assertNotNull($foundGroup->getInformation());
         $this->assertInstanceOf(GroupInformation::class, $foundGroup->getInformation());
-
-        $this->assertNotNull($foundGroup->getInformation()->getGroupCategory());
-        $this->assertInstanceOf(GroupCategory::class, $foundGroup->getInformation()->getGroupCategory());
 
         $this->assertNotNull($foundGroup->getInformation()->getTargetGroup());
         $this->assertInstanceOf(TargetGroup::class, $foundGroup->getInformation()->getTargetGroup());